wifi: mt76: mt7915: move temperature margin check to mt7915_thermal_temp_store()
Originally, we would reduce the 10-degree margin to the restore
temperature, but the user would not be aware of this when setting it.
Moving the margin reduction to the user setting check allows the user to
clearly understand that there is a 10-degree difference between the
restore and trigger temperature.

Signed-off-by: Howard Hsu <howard-yh.hsu@mediatek.com>
Signed-off-by: rany <ranygh@riseup.net>